fix(lib): 修正 SignalLoader 数据读取类型
- 将 fread 函数的 数据类型从 'int16' 修改为 'float32'
- 这个修改解决了数据转换不匹配的问题，确保正确读取数据

+++ lib/SignalLoader.m
@@ -9,7 +9,7 @@ fid = fopen(filename, 'rb');
 fseek(fid, (pointStart - 1) * 8, 'bof');
 
 % 读取数据
-raw = fread(fid, [2, Nread], 'int16');% float32 int16
+raw = fread(fid, [2, Nread], 'float32');% float32 int16
 y = complex(raw(1,:), raw(2,:));
 
 %关闭指针
